blackbox_explainer/backend/main.py
from __future__ import annotations

from dataclasses import replace
import logging
import os
import sys
import time
import uuid
from pathlib import Path
from dotenv import load_dotenv
from fastapi import Depends, FastAPI, HTTPException, Request, status
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field
from sqlalchemy import or_
from sqlmodel import Session, select

PROJECT_ROOT = Path(__file__).resolve().parent.parent
if str(PROJECT_ROOT) not in sys.path:
    sys.path.insert(0, str(PROJECT_ROOT))

from backend.crypto_utils import validate_encryption_setup
from backend.db import get_session, init_db
from backend.dependencies import get_current_user, get_optional_user
from backend.models import User
from backend.schemas import AuthTokenResponse, LoginRequest, MeResponse, SignupRequest
from backend.security import create_access_token, hash_password, verify_password
from config import load_from_env
from explain.pipeline import ExplainerPipeline, build_fallback_result
from llm import create_client


load_dotenv(PROJECT_ROOT / ".env")
load_dotenv(Path(__file__).resolve().parent / ".env")

FOLLOWUP_SYSTEM_PROMPT = """
You are a helpful assistant who explains things clearly to real users.
Rules:
1) Start with a direct answer first.
2) Use plain, natural language instead of formal or academic wording.
3) Explain things like you are talking to a smart user, not writing a paper.
4) If you use a technical term, explain it simply.
5) Explain what the model seems to be focusing on, missing, overstating, or understating when relevant.
6) Use examples only when they make the answer easier to understand.
7) If the follow-up is unrelated to the context, answer directly.
""".strip()

MODEL_OPTIONS = [
    {"value": "tinyllama:latest", "label": "TinyLlama", "description": "Smallest deployment-safe option. Best for getting the hosted explainer to run reliably."},
    {"value": "phi3:mini", "label": "Phi-3 Mini", "description": "Faster and lighter than Llama 3.2, with better quality than TinyLlama when memory allows."},
    {"value": "llama3.2:latest", "label": "Llama 3.2", "description": "Stronger writing quality, but too heavy for many free hosted instances."},
    {"value": "llama3.1:8b", "label": "Llama 3.1 8B", "description": "More capable and detailed. Better when you want stronger reasoning and fuller writeups."},
    {"value": "gpt-oss:120b", "label": "GPT-OSS 120B", "description": "Hosted-scale reasoning model. Best when you want a stronger cloud backend with more depth."},
]

logger = logging.getLogger("bbe.api")
if not logger.handlers:
    logging.basicConfig(level=logging.INFO)

TRIAL_USED_IDS: set[str] = set()
def _cors_origins() -> list[str]:
    raw = os.getenv("CORS_ALLOW_ORIGINS", "").strip()
    if raw:
        if raw == "*":
            return ["*"]
        return [origin.strip() for origin in raw.split(",") if origin.strip()]
    frontend_url = os.getenv("FRONTEND_URL", "").strip()
    if frontend_url:
        return [frontend_url]
    return ["*"]


class ExplainRequest(BaseModel):
    question: str = Field(min_length=1)
    model_answer: str = ""
    context: str = ""
    model: str | None = None


class FollowupRequest(BaseModel):
    question: str = Field(min_length=1)
    model_answer: str = ""
    context: str = ""
    followup: str = Field(min_length=1)
    model: str | None = None


def _load_settings():
    settings = load_from_env()
    base_url = settings.base_url.strip() or "http://127.0.0.1:11434"
    model = settings.model.strip() or "tinyllama:latest"
    return replace(settings, base_url=base_url, model=model)


def _selected_model(requested_model: str | None, default_model: str) -> str:
    allowed_models = {option["value"] for option in MODEL_OPTIONS}
    candidate = (requested_model or default_model or "").strip()
    if not candidate:
        raise HTTPException(status_code=500, detail="Model is not configured on the backend.")
    if candidate not in allowed_models:
        raise HTTPException(status_code=400, detail=f"Unsupported model '{candidate}'.")
    return candidate


def _build_client(model: str):
    settings = _load_settings()
    if not settings.base_url:
        raise HTTPException(status_code=500, detail="Model service URL is not configured on the backend.")
    return create_client(
        base_url=settings.base_url,
        model=model,
        api_key=settings.api_key,
        timeout_seconds=settings.timeout_seconds,
    )


app = FastAPI(title="Black Box Explainer API")
app.add_middleware(
    CORSMiddleware,
    allow_origins=_cors_origins(),
    allow_credentials=False,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.on_event("startup")
def startup_checks() -> None:
    init_db()
    validate_encryption_setup()


@app.middleware("http")
async def operational_logging(request: Request, call_next):
    request_id = str(uuid.uuid4())
    start = time.perf_counter()
    status_code = 500
    try:
        response = await call_next(request)
        status_code = response.status_code
        response.headers["X-Request-ID"] = request_id
        return response
    finally:
        duration_ms = int((time.perf_counter() - start) * 1000)
        logger.info(
            "request_completed request_id=%s method=%s path=%s status=%s duration_ms=%s",
            request_id,
            request.method,
            request.url.path,
            status_code,
            duration_ms,
        )


@app.get("/api/health")
def health():
    settings = _load_settings()
    model = _selected_model(None, settings.model)
    client = _build_client(model)
    ready, status = client.healthcheck()
    return {
        "ok": ready,
        "status": status,
        "serverUrlLocked": True,
        "serverLabel": settings.base_url,
        "selectedModel": model,
        "models": MODEL_OPTIONS,
        "critiquePass": settings.critique_pass,
    }


@app.get("/api/config")
def config():
    settings = _load_settings()
    return {
        "serverUrlLocked": True,
        "serverLabel": settings.base_url,
        "selectedModel": settings.model,
        "models": MODEL_OPTIONS,
    }


@app.post("/api/auth/signup", response_model=MeResponse, status_code=status.HTTP_201_CREATED)
def signup(payload: SignupRequest, session: Session = Depends(get_session)):
    normalized_username = payload.username.strip().lower()
    normalized_email = payload.email.strip().lower()
    existing_username = session.exec(select(User).where(User.username == normalized_username)).first()
    if existing_username:
        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Username already taken.")

    existing = session.exec(select(User).where(User.email == normalized_email)).first()
    if existing:
        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Email already registered.")

    user = User(username=normalized_username, email=normalized_email, password_hash=hash_password(payload.password))
    session.add(user)
    session.commit()
    session.refresh(user)
    return MeResponse(id=user.id or 0, username=user.username, email=user.email)


@app.post("/api/auth/login", response_model=AuthTokenResponse)
def login(payload: LoginRequest, session: Session = Depends(get_session)):
    normalized_identifier = payload.identifier.strip().lower()
    user = session.exec(
        select(User).where(or_(User.email == normalized_identifier, User.username == normalized_identifier))
    ).first()
    if not user or not verify_password(payload.password, user.password_hash):
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid username/email or password.")

    token, expires_in = create_access_token(subject=str(user.id))
    return AuthTokenResponse(access_token=token, expires_in=expires_in)


@app.get("/api/auth/me", response_model=MeResponse)
def me(current_user: User = Depends(get_current_user)):
    return MeResponse(id=current_user.id or 0, username=current_user.username, email=current_user.email)


@app.post("/api/explain")
def explain(
    request: ExplainRequest,
    http_request: Request,
    current_user: User | None = Depends(get_optional_user),
):
    if current_user is None:
        trial_id = http_request.headers.get("X-Trial-Id", "").strip()
        if not trial_id:
            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication required.")
        if trial_id in TRIAL_USED_IDS:
            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Trial used. Please sign in to continue.")
        TRIAL_USED_IDS.add(trial_id)

    settings = _load_settings()
    model = _selected_model(request.model, settings.model)
    client = _build_client(model)
    pipeline = ExplainerPipeline(client)

    try:
        result = pipeline.run(
            question=request.question.strip(),
            model_answer=request.model_answer.strip(),
            context=request.context,
            temperature=settings.temperature,
            max_tokens=settings.max_tokens,
            critique_pass=settings.critique_pass,
        )
    except HTTPException:
        raise
    except Exception:
        result = build_fallback_result(
            question=request.question.strip(),
            model_answer=request.model_answer.strip(),
            context=request.context,
            backend_meta=client.metadata(),
            temperature=settings.temperature,
            max_tokens=settings.max_tokens,
            error_message="Generation failed. Fallback review applied.",
        )

    result["selected_model"] = model
    return result


@app.post("/api/followup")
def followup(request: FollowupRequest, _: User = Depends(get_current_user)):
    settings = _load_settings()
    model = _selected_model(request.model, settings.model)
    client = _build_client(model)

    messages = [
        {"role": "system", "content": FOLLOWUP_SYSTEM_PROMPT},
        {
            "role": "user",
            "content": (
                f"Original question:\n{request.question.strip()}\n\n"
                + (
                    f"Model answer being audited:\n{request.model_answer.strip()}\n\n"
                    if request.model_answer.strip()
                    else ""
                )
                + (
                f"Context:\n{request.context[:2000]}\n\n"
                f"Follow-up question:\n{request.followup.strip()}"
                )
            ),
        },
    ]

    try:
        reply = client.chat(
            messages=messages,
            temperature=settings.temperature,
            max_tokens=min(560, settings.max_tokens),
        )
    except Exception as exc:
        raise HTTPException(status_code=500, detail="The follow-up request failed.") from exc

    return {"reply": reply, "selected_model": model}
