blackbox_explainer/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
requests>=2.31.0
rapidfuzz>=3.9.0
sqlmodel>=0.0.22
sqlalchemy>=2.0.36
alembic>=1.14.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=43.0.3
email-validator>=2.2.0
